+++ policy_to_trajectory/CHANGELOG.rst
@@ -2,9 +2,14 @@
 Changelog for package policy_to_trajectory
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-0.3.1 (2025-05-08)
+0.4.0 (2025-05-15)
 ------------------
 * None
+
+0.3.1 (2025-05-08)
+------------------
+* Updated the LeRobot submodule to the latest version
+* Contributors: Woojin Wie
 
 0.3.0 (2025-04-25)
 ------------------